feat: Crew score Integer에서 Float으로 변경

See merge request s11-bigdata-recom-sub1/S11P21C106!102

+++ project/back/health/src/main/java/com/ssafy/health/domain/battle/dto/response/BattleMatchResponseDto.java
@@ -8,9 +8,9 @@ import lombok.Getter;
 @Builder
 public class BattleMatchResponseDto {
     private String myCrewName;
-    private Integer myCrewScore;
+    private Float myCrewScore;
     private String opponentCrewName;
-    private Integer opponentCrewScore;
+    private Float opponentCrewScore;
     private String exerciseName;
     private BattleStatus battleStatus;
     private Integer dDay;
